+++ modules/apps/commerce/commerce-product-content-search-web/src/main/java/com/liferay/commerce/product/content/search/web/internal/frontend/taglib/form/navigator/RenderSelectionFormNavigatorCategory.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.product.compare.servlet.taglib.ui;
+package com.liferay.commerce.product.content.search.web.internal.frontend.taglib.form.navigator;
 
-import com.liferay.commerce.product.content.web.internal.constants.CPCompareContentConstants;
+import com.liferay.commerce.product.content.search.web.internal.constants.CPSearchResultsConstants;
 import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
@@ -36,12 +36,12 @@ public class RenderSelectionFormNavigatorCategory
 
 	@Override
 	public String getFormNavigatorId() {
-		return CPCompareContentConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
+		return CPSearchResultsConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
 	}
 
 	@Override
 	public String getKey() {
-		return CPCompareContentConstants.CATEGORY_KEY_RENDER_SELECTION;
+		return CPSearchResultsConstants.CATEGORY_KEY_RENDER_SELECTION;
 	}
 
 	@Override
