+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/compare/mini/frontend/taglib/form/navigator/RenderSelectionFormNavigatorCategory.java
@@ -12,13 +12,15 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.product.publisher.servlet.taglib.ui;
+package com.liferay.commerce.product.content.web.internal.product.compare.mini.frontend.taglib.form.navigator;
 
-import com.liferay.commerce.product.content.web.internal.constants.CPPublisherConstants;
+import com.liferay.commerce.product.content.web.internal.constants.CPCompareContentMiniConstants;
 import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
+import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -26,7 +28,7 @@ import org.osgi.service.component.annotations.Component;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	enabled = false, property = "form.navigator.category.order:Integer=20",
+	enabled = false, property = "form.navigator.category.order:Integer=100",
 	service = FormNavigatorCategory.class
 )
 public class RenderSelectionFormNavigatorCategory
@@ -34,17 +36,20 @@ public class RenderSelectionFormNavigatorCategory
 
 	@Override
 	public String getFormNavigatorId() {
-		return CPPublisherConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
+		return CPCompareContentMiniConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
 	}
 
 	@Override
 	public String getKey() {
-		return CPPublisherConstants.CATEGORY_KEY_RENDER_SELECTION;
+		return CPCompareContentMiniConstants.CATEGORY_KEY_RENDER_SELECTION;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "render-selection");
+		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, getClass());
+
+		return LanguageUtil.get(resourceBundle, "render-selection");
 	}
 
 }
